# api_muebles/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Muebles",
        "description": "API REST para la gestión de muebles y categorías de Habita",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:5502/api/v1",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/muebles": {
            "get": {
                "tags": [
                    "Muebles"
                ],
                "summary": "Listar todos los muebles",
                "description": "Endpoint público. Soporta filtros por categoría, precio, color y material, búsqueda por texto, ordenación y paginación.",
                "operationId": "d245c4bc34b577bceb23682eb67d1cc4",
                "parameters": [
                    {
                        "name": "categoria",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "precio_min",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number"
                        },
                        "example": 100
                    },
                    {
                        "name": "precio_max",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number"
                        },
                        "example": 500
                    },
                    {
                        "name": "color",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "blanco"
                    },
                    {
                        "name": "material",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "madera"
                    },
                    {
                        "name": "buscar",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "sofá"
                    },
                    {
                        "name": "orden",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "precio_asc",
                                "precio_desc",
                                "nombre_asc",
                                "nombre_desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de muebles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Sofá Milano"
                                                    },
                                                    "descripcion": {
                                                        "type": "string",
                                                        "example": "Sofá de 3 plazas"
                                                    },
                                                    "precio": {
                                                        "type": "number",
                                                        "example": 299.99
                                                    },
                                                    "stock": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "color": {
                                                        "type": "string",
                                                        "example": "gris"
                                                    },
                                                    "material": {
                                                        "type": "string",
                                                        "example": "tela"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Muebles"
                ],
                "summary": "Crear un nuevo mueble",
                "description": "Requiere token con ability `muebles.crear` (rol Administrador o Gestor).",
                "operationId": "80f97a7997d739eaa9140c8046af09ba",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "precio",
                                    "stock",
                                    "categoria_id"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Sofá Milano"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Sofá de 3 plazas en tela gris"
                                    },
                                    "precio": {
                                        "type": "number",
                                        "example": 299.99
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "gris"
                                    },
                                    "material": {
                                        "type": "string",
                                        "example": "tela"
                                    },
                                    "categoria_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mueble creado correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/muebles/{id}": {
            "get": {
                "tags": [
                    "Muebles"
                ],
                "summary": "Ver un mueble por ID",
                "description": "Endpoint público. Devuelve el mueble con su categoría e imágenes.",
                "operationId": "ab9d946db3bf83042dcc3b36a70825cc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del mueble"
                    },
                    "404": {
                        "description": "Mueble no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Muebles"
                ],
                "summary": "Actualizar un mueble",
                "description": "Requiere token con ability `muebles.editar` (rol Administrador o Gestor).",
                "operationId": "fdcd8678a35383e54e938985cb7cd2ad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Sofá Milano v2"
                                    },
                                    "precio": {
                                        "type": "number",
                                        "example": 349.99
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mueble actualizado correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "404": {
                        "description": "Mueble no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Muebles"
                ],
                "summary": "Eliminar un mueble",
                "description": "Requiere token con ability `muebles.eliminar` (rol Administrador o Gestor).",
                "operationId": "c6717d4fb853dfe7c284fc5af16acf9a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mueble eliminado correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "404": {
                        "description": "Mueble no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token Bearer obtenido desde la API de Usuarios al hacer login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}